Intent: centralize GraphLike protocol for metrics Operators involved: [Coherence, Resonance] Affected invariants: [#9] Key changes: - Move GraphLike protocol to tnfr.types and document its canonical scope - Import GraphLike from tnfr.types across metrics helpers and caches - Keep GraphLike re-exported via metrics.common while updating type hints Expected risks/dissonances: Minimal; ensures a single protocol is reused without altering graph behaviour Metrics: No runtime metrics impacted; existing coherence/SI helpers validated via targeted tests Equivalence map: tnfr.metrics.common.GraphLike -> tnfr.types.GraphLike
