+++ src/tnfr/metrics/trig_cache.py
@@ -14,8 +14,7 @@ from ..alias import get_attr
 from ..constants import get_aliases
 from ..helpers import edge_version_cache
 from ..import_utils import get_numpy
-
-from .common import GraphLike
+from ..types import GraphLike
 
 ALIAS_THETA = get_aliases("THETA")
