+++ src/tnfr/types.py
@@ -1,11 +1,31 @@
-"""Type definitions."""
+"""Type definitions and protocols shared across the engine."""
 
 from __future__ import annotations
+
 from dataclasses import dataclass, field
 from enum import Enum
-from typing import Any
+from typing import Any, Iterable, Protocol
 
-__all__ = ("NodeState", "Glyph")
+__all__ = ("GraphLike", "NodeState", "Glyph")
+
+
+class GraphLike(Protocol):
+    """Protocol for graph objects used throughout TNFR metrics.
+
+    The metrics helpers assume a single coherent graph interface so that
+    coherence, resonance and derived indicators read/write data through the
+    same structural access points.
+    """
+
+    graph: dict[str, Any]
+
+    def nodes(self, data: bool = ...) -> Iterable[Any]: ...
+
+    def number_of_nodes(self) -> int: ...
+
+    def neighbors(self, n: Any) -> Iterable[Any]: ...
+
+    def __iter__(self) -> Iterable[Any]: ...
 
 
 @dataclass(slots=True)
